implementation/requirements.txt
tqdm~=4.67.1
matplotlib~=3.10.0

--extra-index-url https://download.pytorch.org/whl/rocm6.0
torch~=2.6.0
torchvision~=0.21.0
torchaudio==2.3.1
torchmetrics~=1.6.1
pyyaml~=6.0.2
numpy~=2.2.2
pillow~=11.1.0